refactor(admin): verify JWKS tokens with PyJWT, add EC support
The admin runs in .admin-venv, so JWKS verification can use a real crypto
library instead of hand-rolled RSA. Add pyjwt[crypto] to the admin extra
and reimplement the verifier on PyJWT's PyJWKClient, which now accepts
RSA, EC, and EdDSA keys (HS*/none are rejected so a public key can't be
replayed as an HMAC secret) and handles key fetch/cache/rotation.

Verification moves to admin/backend/ (jwks.py + decode_session_token in
auth.py) so the pilot core stays dependency-free; the pure-Python
pilot/commands/jwks.py is removed.
