+++ pilot/managers/admin_env_manager.py
@@ -73,7 +73,7 @@ class AdminEnvManager:
     def _read_admin_deps(self) -> list[str]:
         pyproject = self.venv_path.parent / "pyproject.toml"
         if not pyproject.exists():
-            return ["flask>=3.0", "psutil>=5.9", "pymysql>=1.1", "gunicorn>=21.2"]
+            return ["flask>=3.0", "psutil>=5.9", "pymysql>=1.1", "gunicorn>=21.2", "pyjwt[crypto]>=2.8"]
         with open(pyproject, "rb") as f:
             data = tomllib.load(f)
         return data.get("project", {}).get("optional-dependencies", {}).get("admin")
