+++ pyproject.toml
@@ -12,7 +12,7 @@ dependencies = []
 bench = "pilot.cli:main"
 
 [project.optional-dependencies]
-admin = ["flask>=3.0", "psutil>=5.9", "pymysql>=1.1", "gunicorn>=21.2", "boto3==1.43.39", "psycopg2-binary>=2.9"]
+admin = ["flask>=3.0", "psutil>=5.9", "pymysql>=1.1", "gunicorn>=21.2", "boto3==1.43.39", "psycopg2-binary>=2.9", "pyjwt[crypto]>=2.8"]
 test = ["pytest>=8.0", "pytest-cov>=5.0", "requests>=2.31"]
 # Browser-driven admin-UI lifecycle tests (tests/e2e). pytest-playwright pulls
 # in pytest + playwright; run `playwright install chromium` after installing.
